Introduce Model Hashes
Not really using model hashes for anything yet, but a bunch of things we want to do rely on them so let's get them in there!

- Move functions out of CLI module [Resolves #68]
- Generate model hashes [Resolves #67]
- Add new CLI function to start prototyping modeling runs that are reproducible and save stats to a database. The old one will be maintained for legacy's sake but not actively developed.
- Raise error if seed is set but PYTHONHASHSEED is not [Resolves #54]
- Add reproducibility test to test both model hashing and reproducible behavior
